+++ client/src/core/AppAuthenticator/appAuthenticator.hooks.ts
@@ -1,7 +1,6 @@
 import { useState, useEffect } from 'react';
 
-import * as firebaseui from 'firebaseui';
-import { getAuth, User, EmailAuthProvider, GoogleAuthProvider } from 'firebase/auth';
+import { getAuth, User } from 'firebase/auth';
 
 type AuthState = 'loggedIn' | 'loggedOut' | 'loading';
 
@@ -12,18 +11,6 @@ const useAuthentication = (): {
   const [authState, setAuthState] = useState<AuthState>('loading');
 
   const auth = getAuth();
-  useEffect(() => {
-    if (authState !== 'loggedOut') {
-      return;
-    }
-    const ui = firebaseui.auth.AuthUI.getInstance() || new firebaseui.auth.AuthUI(auth);
-
-    const authConfig = {
-      signInOptions: [EmailAuthProvider.PROVIDER_ID, GoogleAuthProvider.PROVIDER_ID],
-      signInFlow: 'popup',
-    };
-    ui.start('#firebaseui-auth-container', authConfig);
-  }, [authState]);
 
   useEffect(() => {
     const authListener = auth.onAuthStateChanged((result) => {
