+++ client/src/pages/Login/index.ts
@@ -1,0 +1,3 @@
+import Login from './Login';
+
+export default Login;
