+++ client/src/pages/index.ts
@@ -2,3 +2,4 @@ export { default as Accounts } from './Accounts';
 export { default as Tags } from './Tags';
 export { default as Transactions } from './Transactions';
 export { default as Landing } from './Landing';
+export { default as Login } from './Login';
